+++ mods/iteminteractions/1.21.10/Common/src/main/java/fuzs/iteminteractions/api/v1/provider/ItemContentsProvider.java
@@ -52,8 +52,9 @@ public interface ItemContentsProvider {
      * Codec that includes a list of supported items.
      */
     Codec<Map.Entry<HolderSet<Item>, ItemContentsProvider>> WITH_ITEMS_CODEC = RecordCodecBuilder.create(instance -> {
-        return instance.group(Ingredient.NON_AIR_HOLDER_SET_CODEC.optionalFieldOf("supported_items", HolderSet.empty())
-                .forGetter(Map.Entry::getKey), CODEC.forGetter(Map.Entry::getValue)).apply(instance, Map::entry);
+        return instance.group(Ingredient.NON_AIR_HOLDER_SET_CODEC.lenientOptionalFieldOf("supported_items",
+                        HolderSet.empty()).forGetter(Map.Entry::getKey), CODEC.forGetter(Map.Entry::getValue))
+                .apply(instance, Map::entry);
     });
     /**
      * Stream codec that additionally to the provider itself also includes the provider type.