+++ mods/iteminteractions/1.21.10/CHANGELOG.md
@@ -5,6 +5,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [v21.10.1-1.21.10] - 2025-11-18
+
+### Changed
+
+- The `supported_items` field is now optional and may be empty when specified
+- Add a few more helper methods in `AbstractItemContentsProvider`
+
 ## [v21.10.0-1.21.10] - 2025-10-08
 
 ### Changed